Move tests to a folder
